Replaced vivid `.prelude` banner with subtler text.
The intensity of the reds was distracting from content, and distracting from me improving the page styling more generally. Despite several steps over the last week+ I couldn't find colours or patterns that worked, so I decided to find a different design space.

The result is in this commit.

* The navigation bar is changed to red-purple, as I prefer it as a calmer colour choice to the previous orange-red. The colour is somewhat lighter to compensate for...
* The big banner that contains post titles, `.prelude`, has been replaced with something coloured the same as the `body`. Extended text in these areas (e.g., homepage) has been removed. The heading colour has been set to the same red-purple as the navigation bar.
* Post pages now use the same date style as post lists.
* Posts in general now have a yellow border across their top, in line with its removal from the header+banner.

I'm rather happy with the long-term sustainability of this change.