Redeploy the site every day
Jekyll only renders posts whose date is not in the future. Future
posts will still be visible in GitHub, but not on my website. I'm OK
with the compromise for now.

Having Github Actions trigger every morning means that future posts
will appear on the website on the proper day.